Fixed navigation and functionality of addon services
Previously due to the fact that the parent state was stored *only*
in the bazarr class instance, any other call to bazarr would cause
the state to be invalidated by the new call.
Therefore any other use that e.g. is at the same time trying to
download something with bazarr will have issues due to the change
of the underlying series/movie without their knowledge.

This can not happen from now on, however the current implementation is
still rather limiting, as most of the parents state is not accessible.

To properly solve this, it would be best to give the state key to the
addon and then the addon can request the state data whenever it wants
and therefore get all relevant data itself.